[이용훈] 2024.05.17 commit5
[추가]
1. 테마 삭제 시 해당하는 room이 있으면 삭제 불가

+++ src/main/webapp/emp/deleteThemeAction.jsp
@@ -1,5 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ page import="beeNb.dao.ThemeDAO" %>
+<%@ page import="beeNb.dao.RoomDAO" %>
 <%@page import="java.net.URLEncoder"%>
 <%
 	System.out.println("=====deleteThemeAction.jsp=====");
@@ -8,6 +9,17 @@
 	String roomTheme = request.getParameter("roomTheme");
 	// roomTheme 디버깅
 	System.out.println("roomTheme : " + roomTheme);
+	
+	// 해당 테마를 사용하고있는 room이 있는지 조회
+	// return : boolean (있으면 true, 없으면 false)
+	boolean thereIsRoomTheme = RoomDAO.thereIsRoomTheme(roomTheme);
+	System.out.println("thereIsRoomTheme : " + thereIsRoomTheme);
+	if(thereIsRoomTheme==true){
+		System.out.println("사용중인 테마");
+		String msg = URLEncoder.encode("사용중인 테마입니다. 해당 숙소를 삭제한 후 시도해주세요.", "UTF-8");
+		response.sendRedirect("/BeeNb/emp/themeList.jsp?msg="+msg);
+		return;
+	}
 	
 	// 테마 삭제하는 메서드
 	int deleteTheme = ThemeDAO.deleteTheme(roomTheme);
